refactor: /simplify pass on #418 — split CapabilityEnforcer, hoist redundant normalization
- CapabilityEnforcer.cs (grown to ~390 lines through five review-driven
  hardening rounds) split by concern into CapabilityEnforcer.PathScoping.cs
  and CapabilityEnforcer.HostScoping.cs, matching this codebase's existing
  large-class convention (PlanExecutor.Scheduling.cs/Recovery.cs,
  ToolInvocationGovernor). Pure reorganization, no behavior change.
- EnforcePathScoping/EnforceHostScoping now canonicalize each configured
  DeniedPaths/AllowedPaths/DeniedHosts/AllowedHosts entry once per
  EnforceAsync call instead of once per (requested value x configured
  entry) pair inside the comparison loop -- real savings when a
  canonicalizer is registered, since each entry costs an actual filesystem
  stat/symlink resolution, not just a string comparison.
- FileSystemTool.ResourceParametersByOperation now shares one dictionary
  instance across all five operations instead of allocating an identical
  one per operation.

All existing tests pass unchanged, confirming the refactor is behavior-
preserving.
