+++ src/Content/Infrastructure/Infrastructure.AI/Tools/FileSystemTool.cs
@@ -32,6 +32,10 @@ public sealed class FileSystemTool : ITool
     private static readonly IReadOnlyList<string> Operations =
         ["read", "write", "list", "search", "exists"];
 
+    /// <summary>The single shared declaration every operation maps to — see <see cref="ResourceParametersByOperation"/>.</summary>
+    private static readonly IReadOnlyDictionary<string, ResourceParameterKind> PathOnly =
+        new Dictionary<string, ResourceParameterKind> { ["path"] = ResourceParameterKind.Path };
+
     private readonly IFileSystemService _fileSystem;
 
     /// <summary>
@@ -76,10 +80,7 @@ public sealed class FileSystemTool : ITool
     /// itself reads every operation's <c>path</c> argument via <c>GetRequiredString</c>.
     /// </remarks>
     public IReadOnlyDictionary<string, IReadOnlyDictionary<string, ResourceParameterKind>> ResourceParametersByOperation { get; } =
-        Operations.ToDictionary(
-            op => op,
-            IReadOnlyDictionary<string, ResourceParameterKind> (op) =>
-                new Dictionary<string, ResourceParameterKind> { ["path"] = ResourceParameterKind.Path });
+        Operations.ToDictionary(op => op, _ => PathOnly);
 
     /// <inheritdoc />
     public async Task<ToolResult> ExecuteAsync(
